+++ lib/src/_computation.dart
@@ -13,11 +13,13 @@ class Computation {
   final void Function(Computation) _progress;
 
   FutureOr? _pending;
-  dynamic _result;
+  Object? _result;
   Object? error;
   StackTrace? stackTrace;
 
-  dynamic get result {
+  final _completer = Completer<void>();
+
+  Object? get result {
     if (!_completer.isCompleted) {
       // too early: not completed yet
       throw StateError('Asynchronous computation is still pending.');
@@ -30,21 +32,11 @@ class Computation {
     }
   }
 
-  final _completer = Completer<void>();
-
   bool get isSuccess => _completer.isCompleted && (error == null);
   bool get isFailure => _completer.isCompleted && (error != null);
 
-  void compute(Results results) {
+  void run(Results results) {
     // completion
-    void $complete(dynamic value) {
-      if (!_completer.isCompleted) {
-        _result = value;
-        _completer.complete();
-        _progress(this);
-      }
-    }
-
     void $completeError(Object error, [StackTrace? stackTrace]) {
       if (!_completer.isCompleted) {
         this.error = error;
@@ -54,36 +46,38 @@ class Computation {
       }
     }
 
+    void $complete(FutureOr value) {
+      if (value case Future f) {
+        f.then($complete, onError: $completeError);
+      } else if (!_completer.isCompleted) {
+        _result = value;
+        _completer.complete();
+        _progress(this);
+      }
+    }
+
     // start computation
     try {
-      final res = _pending = _start(results);
-      if (res case Future f) {
-        // async route
-        f.then($complete, onError: $completeError);
-      } else {
-        // sync route (no error)
-        $complete(res);
-      }
+      $complete(_pending = _start(results));
     } catch (err, st) {
-      // sync route (error)
       $completeError(err, st);
     }
   }
 
-  Future get future {
+  Future<Object?> get future {
     final pending = _pending;
-    return (pending is Future) ? pending : Future.value(pending);
+    return (pending is Future<Object?>) ? pending : Future.value(pending);
   }
 
   void cleanup() {
     if (_cleanup == null || error != null) {
-      // completed with an error: nothing to do
+      // no cleanup provided, or completed with an error: nothing to do
     } else if (!_completer.isCompleted) {
       // not completed yet: ensure cleanup on completion
       _completer.future.whenComplete(cleanup);
-    } else if (_result != null) {
+    } else if (_result case Object res) {
       // completed and non-null result: cleanup
-      _cleanup(_result);
+      _cleanup(res);
     }
   }
 }